requirements.txt
git+ssh://git@g.krk.ru/orion-crm/libs.rabbitmq.git
git+ssh://git@g.krk.ru/crm-erp/crm-erp-models@staging
git+ssh://git@g.krk.ru/orion-crm/libs.logs-manager.git
git+ssh://git@g.krk.ru/orion-crm/libs.database.git@v1.0.1


six==1.16.0
numpy==1.21.2
Pillow==8.0.1
PyYAML==5.3.1
imutils==0.5.4
aiohttp==3.7.2
gunicorn==20.1.0
PyWavelets==1.1.1
matplotlib==3.4.3
sentry-sdk==0.19.1
pytesseract==0.3.8
python-dateutil==2.8.2
opencv-python==4.5.4.58
